Move model default value apis to model manager facade
