+++ apiserver/modelconfig/modelconfig_test.go
@@ -42,20 +42,6 @@ func (s *modelconfigSuite) SetUpTest(c *gc.C) {
 			"ftp-proxy":       {"http://proxy", "model"},
 			"authorized-keys": {testing.FakeAuthKeys, "model"},
 		},
-		cfgDefaults: config.ModelDefaultAttributes{
-			"attr": config.AttributeDefaultValues{
-				Default:    "",
-				Controller: "val",
-				Regions: []config.RegionDefaultValue{config.RegionDefaultValue{
-					Name:  "dummy",
-					Value: "val++"}}},
-			"attr2": config.AttributeDefaultValues{
-				Controller: "val3",
-				Default:    "val2",
-				Regions: []config.RegionDefaultValue{config.RegionDefaultValue{
-					Name:  "left",
-					Value: "spam"}}},
-		},
 	}
 	var err error
 	s.api, err = modelconfig.NewModelConfigAPI(s.backend, &s.authorizer)
@@ -167,164 +153,15 @@ func (s *modelconfigSuite) TestModelUnsetMissing(c *gc.C) {
 	c.Assert(err, jc.ErrorIsNil)
 }
 
-func (s *modelconfigSuite) TestModelDefaults(c *gc.C) {
-	result, err := s.api.ModelDefaults()
-	c.Assert(err, jc.ErrorIsNil)
-	expectedValues := map[string]params.ModelDefaults{
-		"attr": {
-			Controller: "val",
-			Default:    "",
-			Regions: []params.RegionDefaults{{
-				RegionName: "dummy",
-				Value:      "val++"}}},
-		"attr2": {
-			Controller: "val3",
-			Default:    "val2",
-			Regions: []params.RegionDefaults{{
-				RegionName: "left",
-				Value:      "spam"}}},
-	}
-	c.Assert(result.Config, jc.DeepEquals, expectedValues)
-}
-
-func (s *modelconfigSuite) TestSetModelDefaults(c *gc.C) {
-	params := params.SetModelDefaults{
-		Config: []params.ModelDefaultValues{{
-			Config: map[string]interface{}{
-				"attr3": "val3",
-				"attr4": "val4"},
-		}}}
-	result, err := s.api.SetModelDefaults(params)
-	c.Assert(err, jc.ErrorIsNil)
-	c.Assert(result.OneError(), jc.ErrorIsNil)
-	c.Assert(s.backend.cfgDefaults, jc.DeepEquals, config.ModelDefaultAttributes{
-		"attr": {
-			Controller: "val",
-			Default:    "",
-			Regions: []config.RegionDefaultValue{{
-				Name:  "dummy",
-				Value: "val++"}}},
-		"attr2": {
-			Controller: "val3",
-			Default:    "val2",
-			Regions: []config.RegionDefaultValue{{
-				Name:  "left",
-				Value: "spam"}}},
-		"attr3": {Controller: "val3"},
-		"attr4": {Controller: "val4"},
-	})
-}
-
-func (s *modelconfigSuite) TestBlockChangesSetModelDefaults(c *gc.C) {
-	s.blockAllChanges(c, "TestBlockChangesSetModelDefaults")
-	_, err := s.api.SetModelDefaults(params.SetModelDefaults{})
-	s.assertBlocked(c, err, "TestBlockChangesSetModelDefaults")
-}
-
-func (s *modelconfigSuite) TestUnsetModelDefaults(c *gc.C) {
-	args := params.UnsetModelDefaults{
-		Keys: []params.ModelUnsetKeys{{
-			Keys: []string{"attr"},
-		}}}
-	result, err := s.api.UnsetModelDefaults(args)
-	c.Assert(err, jc.ErrorIsNil)
-	c.Assert(result.OneError(), jc.ErrorIsNil)
-	c.Assert(s.backend.cfgDefaults, jc.DeepEquals, config.ModelDefaultAttributes{
-		"attr2": {
-			Controller: "val3",
-			Default:    "val2",
-			Regions: []config.RegionDefaultValue{{
-				Name:  "left",
-				Value: "spam"}}},
-	})
-}
-
-func (s *modelconfigSuite) TestBlockUnsetModelDefaults(c *gc.C) {
-	s.blockAllChanges(c, "TestBlockUnsetModelDefaults")
-	args := params.UnsetModelDefaults{
-		Keys: []params.ModelUnsetKeys{{
-			Keys: []string{"abc"},
-		}}}
-	_, err := s.api.UnsetModelDefaults(args)
-	s.assertBlocked(c, err, "TestBlockUnsetModelDefaults")
-}
-
-func (s *modelconfigSuite) TestUnsetModelDefaultsMissing(c *gc.C) {
-	// It's okay to unset a non-existent attribute.
-	args := params.UnsetModelDefaults{
-		Keys: []params.ModelUnsetKeys{{
-			Keys: []string{"not there"},
-		}}}
-	result, err := s.api.UnsetModelDefaults(args)
-	c.Assert(err, jc.ErrorIsNil)
-	c.Assert(result.OneError(), jc.ErrorIsNil)
-}
-
-func (s *modelconfigSuite) TestModelDefaultsAsNormalUser(c *gc.C) {
-	mc, err := modelconfig.NewModelConfigAPI(s.backend, apiservertesting.FakeAuthorizer{
-		Tag: names.NewUserTag("charlie@local")})
-	c.Assert(err, jc.ErrorIsNil)
-	got, err := mc.ModelDefaults()
-	c.Assert(err, gc.ErrorMatches, "permission denied")
-	c.Assert(got, gc.DeepEquals, params.ModelDefaultsResult{})
-}
-
-func (s *modelconfigSuite) TestSetModelDefaultsAsNormalUser(c *gc.C) {
-	mc, err := modelconfig.NewModelConfigAPI(s.backend, apiservertesting.FakeAuthorizer{
-		Tag: names.NewUserTag("charlie@local")})
-	c.Assert(err, jc.ErrorIsNil)
-	got, err := mc.SetModelDefaults(params.SetModelDefaults{
-		Config: []params.ModelDefaultValues{{
-			Config: map[string]interface{}{
-				"ftp-proxy": "http://charlie",
-			}}}})
-	c.Assert(err, jc.ErrorIsNil)
-	c.Assert(got, jc.DeepEquals, params.ErrorResults{
-		Results: []params.ErrorResult{
-			params.ErrorResult{
-				Error: &params.Error{
-					Message: "permission denied",
-					Code:    "unauthorized access"}}}})
-
-	// Make sure it didn't change.
-	cfg, err := s.api.ModelGet()
-	c.Assert(err, jc.ErrorIsNil)
-	c.Assert(cfg.Config["ftp-proxy"].Value.(string), jc.DeepEquals, "http://proxy")
-}
-
-func (s *modelconfigSuite) TestUnsetModelDefaultsAsNormalUser(c *gc.C) {
-	mc, err := modelconfig.NewModelConfigAPI(s.backend, apiservertesting.FakeAuthorizer{
-		Tag: names.NewUserTag("charlie@local")})
-	c.Assert(err, jc.ErrorIsNil)
-	got, err := mc.UnsetModelDefaults(params.UnsetModelDefaults{
-		Keys: []params.ModelUnsetKeys{{
-			Keys: []string{"ftp-proxy"}}}})
-	c.Assert(err, gc.ErrorMatches, "permission denied")
-	c.Assert(got, gc.DeepEquals, params.ErrorResults{
-		Results: []params.ErrorResult{
-			params.ErrorResult{
-				Error: nil}}})
-
-	// Make sure it didn't change.
-	cfg, err := s.api.ModelGet()
-	c.Assert(err, jc.ErrorIsNil)
-	c.Assert(cfg.Config["ftp-proxy"].Value.(string), jc.DeepEquals, "http://proxy")
-}
-
 type mockBackend struct {
-	cfg         config.ConfigValues
-	cfgDefaults config.ModelDefaultAttributes
-	old         *config.Config
-	b           state.BlockType
-	msg         string
+	cfg config.ConfigValues
+	old *config.Config
+	b   state.BlockType
+	msg string
 }
 
 func (m *mockBackend) ModelConfigValues() (config.ConfigValues, error) {
 	return m.cfg, nil
-}
-
-func (m *mockBackend) ModelConfigDefaultValues() (config.ModelDefaultAttributes, error) {
-	return m.cfgDefaults, nil
 }
 
 func (m *mockBackend) UpdateModelConfig(update map[string]interface{}, remove []string, validate state.ValidateConfigFunc) error {
@@ -339,16 +176,6 @@ func (m *mockBackend) UpdateModelConfig(update map[string]interface{}, remove []
 	}
 	for _, n := range remove {
 		delete(m.cfg, n)
-	}
-	return nil
-}
-
-func (m *mockBackend) UpdateModelConfigDefaultValues(update map[string]interface{}, remove []string) error {
-	for k, v := range update {
-		m.cfgDefaults[k] = config.AttributeDefaultValues{Controller: v}
-	}
-	for _, n := range remove {
-		delete(m.cfgDefaults, n)
 	}
 	return nil
 }
